Fix stuck merge
